Relocate & rename discord service
