+++ src/common/interfaces.ts
@@ -12,12 +12,3 @@ export interface IProjectList {
   projects: IProject[];
 }
 
-
-export interface IDiscordAuthResponse {
-  "access_token": string;
-  "token_type": "Bearer";
-  "expires_in": number;
-  "expires_at"?: number;
-  "refresh_token": string;
-  "scope": string;
-}
