+++ src/common/services/projects.ts
@@ -1,0 +1,1 @@
+export default ""
