+++ src/views/signin.tsx
@@ -1,6 +1,6 @@
 import { Text, Stack, Label, Spinner } from "office-ui-fabric-react";
 import React, { useState, useEffect } from "react";
-import { AuthData } from "../common/discordService";
+import { AuthData } from "../common/services/discord";
 
 export const Signin: React.FC<any> = () => {
